+++ src/components/ShoppingCart/useCartDispatch.ts
@@ -13,6 +13,7 @@ import {
   getCartItems,
   clearselectedItems,
   moveOrder,
+  selectedOrder,
 } from "../../slice/cartSlice";
 
 export const useCartDispatch = () => {
@@ -21,6 +22,7 @@ export const useCartDispatch = () => {
   const selectedItems = useAppSelector((state) => state.cart.selectedItems);
   const navigate = useNavigate();
 
+  //장바구니 전체 선택, 혹은 현재 전체 선택되어있다면 전체 해제합니다.
   const handleSelectAll = (isChecked: boolean) => {
     if (isChecked) {
       dispatch(selectAllItems());
@@ -29,6 +31,7 @@ export const useCartDispatch = () => {
     }
   };
 
+  //해당 상품을 선택하거나 선택 취소합니다.
   const handleItemSelect = (itemToSelect: CartItem) => {
     if (selectedItems.some((item) => item.id === itemToSelect.id)) {
       dispatch(deselectItem(itemToSelect.id));
@@ -36,43 +39,48 @@ export const useCartDispatch = () => {
       dispatch(selectItem(itemToSelect));
     }
   };
-
+  //선택된 값들을 초기화해줍니다. 장바구니에 처음 진입할때만 사용됩니다.
   const selectClear = () => {
     dispatch(clearselectedItems());
   };
-
+  //서버로부터 내 장바구니 데이터를 받아옵니다.
   const getCartlisting = () => {
     dispatch(getCartItems());
   };
-
+  //단일 상품을 삭제합니다.
   const handleDelete = (id: number) => {
     dispatch(removeItem(id));
   };
-
+  //장바구니의 상품들을 전체 삭제합니다.
   const handleDeleteAll = () => {
     dispatch(removeAll());
   };
-
+  //장바구니에서 선택된 상품들만 삭제합니다.
   const handleDeleteSelected = () => {
     dispatch(removeSelected());
   };
-
+  //현재 장바구니의 상품 수량을 1씩 증가시킵니다.
   const plusQuantity = (id: number) => {
     const item = items.find((item) => item.id === id);
     if (item && item.amount < item.productStock) {
       dispatch(incrementQuantity(id));
     }
   };
-
+  //현재 장바구니의 상품 수량을 1씩 감소시킵니다.
   const minusQuantity = (id: number) => {
     const item = items.find((item) => item.id === id);
     if (item && item.amount > 1) {
       dispatch(decrementQuantity(id));
     }
   };
-
+  //현재 장바구니에서 선택된 상품들을 결제페이지에서 조회할수있게 값을 보내주고, 결제페이지로 이동합니다.
   const goOrder = () => {
     dispatch(moveOrder());
+    navigate("/order");
+  };
+  //단일 상품 구매입니다.
+  const SelectgoOrder = (id: number) => {
+    dispatch(selectedOrder(id));
     navigate("/order");
   };
 
@@ -89,5 +97,6 @@ export const useCartDispatch = () => {
     getCartlisting,
     selectClear,
     goOrder,
+    SelectgoOrder,
   };
 };
